implementing Linux commands in C language
Design and implement program (task11prog.c its executable named task11prog) to create directories and files, and/or to navigate the directories as shown below. Further provide Makefile to compile the program.

Implement the following steps of the tasks to be performed. 
1. Change the directory (chdir) to be: project2
2. Get current working directory (getçwd) to be output
3. Get the directory listing of the current working directory and output the
4. Create a directory (mkdir) dir0
5. Change the directory (chdir) to be: dir0 
6. Create a directory (mkdir): dir1
7. Change the directory (chdir) to be: dir1
8. Create a file: filel.txt 
9. Create a file: file2. Txt
10. Get the directory listing of the current working directory and output the listing 11. Change the directory (chdir) to the parent directory ".")
12. Create a directory (mkdir): dir2
// Change the directory (chdir): dir2 
13. Create a file: file3.txt
14. Create a file: file4. Txt
15. Get the directory listing of the current working directory and output the listing 16. Change the directory (chdir) to the parent directory ("..")
17. // The current working directory is: dir0
18. Create a file (symbolic link to dir2/file4.txt): link5.txt
19. Create a file: file. txt
20. Get the directory listing of the current working directory and output the listing

For the program output, provide a heading for each step in your program and to the console output so that you can see the execution of your program step by step. Each heading should include a brief message with a time-of-day timestamp (e.g., " ** Step 1. Change Directory to be project2 - Feb 03, 2021 05:20:35 PM **").